make mock-socket dependency
