+++ package.json
@@ -12,7 +12,8 @@
     "pino": "8.6.1",
     "pino-pretty": "9.1.0",
     "prom-client": "13.2.0",
-    "ws": "8.9.0"
+    "ws": "8.9.0",
+    "mock-socket": "9.1.5"
   },
   "scripts": {
     "build": "mkdir -p ./dist/src && cp package.json dist/src && cp README.md dist/src && tsc",
@@ -42,7 +43,6 @@
     "eslint": "8.39.0",
     "eslint-config-prettier": "8.8.0",
     "eslint-plugin-tsdoc": "0.2.17",
-    "mock-socket": "9.1.5",
     "mocksse": "1.0.4",
     "prettier": "2.8.7",
     "ts-node": "10.9.1",
