+++ src/util/testing-utils.ts
@@ -8,9 +8,11 @@ import { Adapter, AdapterDependencies } from '../adapter'
 import { Cache, LocalCache } from '../cache'
 import { ResponseCache } from '../cache/response'
 import { EmptyCustomSettings, SettingsDefinitionMap } from '../config'
-import { Transport, TransportDependencies, TransportGenerics } from '../transports'
+import { Transport, TransportDependencies, TransportGenerics, WebSocketClassProvider } from '../transports'
 import { AdapterRequest, AdapterResponse, PartialAdapterResponse, sleep } from './index'
 import { EmptyInputParameters, TypeFromDefinition } from '../validation/input-params'
+import { WebSocket } from 'mock-socket'
+export { WebSocket as MockWebsocket, Server as MockWebsocketServer } from 'mock-socket'
 
 export type NopTransportTypes = {
   Parameters: EmptyInputParameters
@@ -443,5 +445,35 @@ export function setEnvVariables(envVariables: NodeJS.ProcessEnv): void {
   for (const key in envVariables) {
     process.env[key] = envVariables[key]
   }
+}
+
+/**
+ * Sets the mocked websocket instance in the provided provider class.
+ * We need this here, because the tests will connect using their instance of WebSocketClassProvider;
+ * fetching from this library to the \@chainlink/ea-bootstrap package would access _another_ instance
+ * of the same constructor. Although it should be a singleton, dependencies are different so that
+ * means that the static classes themselves are also different.
+ *
+ * @param provider - singleton WebSocketClassProvider
+ */
+export const mockWebSocketProvider = (provider: typeof WebSocketClassProvider): void => {
+  // Extend mock WebSocket class to bypass protocol headers error
+  class MockWebSocket extends WebSocket {
+    constructor(url: string, protocol: string | string[] | Record<string, string> | undefined) {
+      super(url, protocol instanceof Object ? undefined : protocol)
+    }
+    // This is part of the 'ws' node library but not the common interface, but it's used in our WS transport
+    removeAllListeners() {
+      for (const eventType in this.listeners) {
+        // We have to manually check because the mock-socket library shares this instance, and adds the server listeners to the same obj
+        if (!eventType.startsWith('server')) {
+          delete this.listeners[eventType]
+        }
+      }
+    }
+  }
+
+  // Need to disable typing, the mock-socket impl does not implement the ws interface fully
+  provider.set(MockWebSocket as any) // eslint-disable-line @typescript-eslint/no-explicit-any
 }
 /* c8 ignore stop */ // eslint-disable-line
