+++ test/metrics/ws-metrics.test.ts
@@ -5,8 +5,7 @@ import { Adapter, AdapterEndpoint } from '../../src/adapter'
 import { AdapterConfig, EmptyCustomSettings } from '../../src/config'
 import { WebSocketClassProvider, WebSocketTransport } from '../../src/transports'
 import { InputParameters } from '../../src/validation'
-import { TestAdapter } from '../../src/util/testing-utils'
-import { mockWebSocketProvider } from '../utl'
+import { TestAdapter, mockWebSocketProvider } from '../../src/util/testing-utils'
 
 export const test = untypedTest as TestFn<{
   adapterEndpoint: AdapterEndpoint<WebSocketEndpointTypes>
