+++ test/transports/routing.test.ts
@@ -17,8 +17,7 @@ import {
   WebSocketTransport,
 } from '../../src/transports'
 import { InputParameters } from '../../src/validation'
-import { TestAdapter } from '../../src/util/testing-utils'
-import { mockWebSocketProvider } from '../utl'
+import { TestAdapter, mockWebSocketProvider} from '../../src/util/testing-utils'
 
 const test = untypedTest as TestFn<{
   testAdapter: TestAdapter<SettingsDefinitionFromConfig<typeof adapterConfig>>
